Disable on simulation pages
I changed the declarativeContent to use a regex that matches /ide but not /ide/simulation.
The one caveat here is chrome uses re2, which doesn't allow negative lookaheads. As a work
around, I only check the first 4 letters (simu). With a million documents the odds of that
being part of a file id is less than 1%, so I think we're safe.