Separate commit message into title and body
I made CommitMessage a separate component, which is *so* much cleaner. I think I'm going to modularize basically everything in App.jsx's render function very soon.
- CommitMessage contains a title textarea and a body textarea.
- The title has a maxlength of 50 and disallows newline characters.
- The component joins the title and body with a newline character when it passes it to the App component
- The textarea is dramatically larger to encourage good documentation

Minor side change:
I redid the declarativeContent code that checks whether to enable the page action on the current page. Now it checks
	NOT /ide/ *10 characters* /
rather than
	NOT /ide/simul
because the latter could theoretically occur in a valid document ID, whereas the former never could. Not a big change, but it was bothering me