Change merge to use ace-diff
No longer uses the merge feature of the GitHub API ever.
Instead, ace-diff is put on the page and yous can make a commit with two parents (a merge)
I removed diff because it's too similar to merge and people would get confused and commit
things that needed two parents with only one.