Add simulation dialog branch menu
There's a new setting in advanced. When you turn it on, the simulation dialog has a new option, next to "select standard player." When clicked it gives you a list of GitHub branches. You can sim against them. I made a few minor changes to a lot of documents to make this work.