Add Log in button
In the interest of making things more intuitive and familiar, there is now a logged-in status region in the upper right. It will say who you are logged in as, and if you are not logged in, it will display a login button. It has not been thoroughly tested, and it violates the OAuth2 good practices of presenting logins in context. I also changed the eventApi query function to assume the repo you are using. Also, I fixed a JSON error in the manifest-example.